// frontend/src/types/turtle.ts
export interface TurtleRecord {
  id: string;
  originalSite: string;
  name: string;
  sex: string;
  dateLastSeen: string;
  lastFoundBy: string;
  locationDescription: string;
  recaptureHistory: string;
  dateFirstFound: string;
  health: string;
  dateLastBehavior: string;
  generalNotes: string;
}